+++ .dialyzer_ignore.exs
@@ -1,0 +1,3 @@
+[
+  {"lib/routex/extension/localize/phoenix/localize_phoenix_routes.ex", :unknown_function}
+]